Rewards are float

+++ simple_continuous.py
@@ -93,9 +93,9 @@ def main():
     parser.add_argument('--n_poison', type=int, default=10)
     parser.add_argument('--n_coop', type=int, default=2)
     parser.add_argument('--n_sensors', type=int, default=30)
-    parser.add_argument('--food_reward', type=int, default=3)
-    parser.add_argument('--poison_reward', type=int, default=-1)
-    parser.add_argument('--encounter_reward', type=int, default=0.05)
+    parser.add_argument('--food_reward', type=float, default=3)
+    parser.add_argument('--poison_reward', type=float, default=-1)
+    parser.add_argument('--encounter_reward', type=float, default=0.05)
 
     parser.add_argument('--policy_hidden_spec', type=str, default=GAE_ARCH)
